Changed to the default firefly page
Now that the default view of Firefly is about to get a little more visibility I am removing some of the development buttons
from the default page. The example dialog, original drop down button that we used for testing, and the help are no longer on
the page.  I reordered the buttons as well.

We will put the help button back when we actually have a help site to go to.

Out default view is:

http://localhost:8080/firefly/

To get at the old view, which is important if you want to look at the example dialog, use the following link:

http://localhost:8080/firefly/firefly-dev.html