Firefly-1225: fixed: multiple issues
  - Firefly-1225: fixed: image not rendering in safari 16
  - Firefly-1226: fixed, sort of: The data products tab no longer show the wrong table title on the third fetch
  - Firefly-1224: fixed chart causing crash
  - Added helpid as a prop to TargetHiPSPanel
  - Fixed: search refinment tool showing wrong radius on repeated uses
  - Fixed: readout only shows on image with mouse on it
  - Fixed: Remember TAP services showing
  - HiPS fetching optimizations.

TLDL;
The problem when a server is down: When we have the file cached and local we still call the URL with the If-Modified-Since header. This allows the server to give a not-modified
(304) response. That way we check to see if the data is the most recent.

This is not really necessary with hips tiles when it is coming from a known HiPS in the HiPSList. These hips tiles will almost never change. If it was a hips repository not on
the hips list then it would be different, the hips tiles might change,  because the hips repository is probably under development.

I am going to make two changes.
  - For tiles from hips in the hipslist, it almost never makes sense to call the server, they are very, very static. You would have to regenerate a hips to have this change.
    So, don't call the server, just return the tile
  - for hipslist, property files, and hips tiles not in a hipslist. It will change if the file is already downloaded, then  make the call with a really small timeout. That way
    if the server is down it will not cause a substantial slowdown and since we cache these file we will probably have something to return.