IRSA-5807: disable the proprietary data warning.
  - This feature was never working right

This issue is easily fixed by removing the message completely. The problem is we put this in during the sofia (2019) work that does have proprietary data. But it never worked right. We just never noticed before. Firefly loads tables in pages, if the table is not fully loaded on the client and you try to download then it will give always give you that message.
In the ticket example we had only loaded 100 of the 136 rows.